+++ lib/save_it/application.ex
@@ -7,26 +7,28 @@ defmodule SaveIt.Application do
 
   @impl true
   def start(_type, _args) do
-    token = Application.fetch_env!(:save_it, :telegram_bot_token) |> require_telegram_bot_token!()
-
     :logger.add_handler(:my_sentry_handler, Sentry.LoggerHandler, %{
       config: %{metadata: [:file, :line]}
     })
 
-    children =
-      if Application.get_env(:save_it, :start_bot?, true) do
-        [
-          ExGram,
-          {SaveIt.Bot, [method: :polling, token: token]}
-        ]
-      else
-        []
-      end
-
     # See https://hexdocs.pm/elixir/Supervisor.html
     # for other strategies and supported options
     opts = [strategy: :one_for_one, name: SaveIt.Supervisor]
-    Supervisor.start_link(children, opts)
+    Supervisor.start_link(children(), opts)
+  end
+
+  @doc false
+  def children do
+    token = Application.fetch_env!(:save_it, :telegram_bot_token) |> require_telegram_bot_token!()
+
+    if Application.get_env(:save_it, :start_bot?, true) do
+      [
+        ExGram,
+        {SaveIt.Bot, [method: SaveIt.TelegramPolling, token: token]}
+      ]
+    else
+      []
+    end
   end
 
   defp require_telegram_bot_token!(token) when is_binary(token) do